DEV-6109 Support tables with multiple tgroup elements
This also extends the table (tgroup) selector part to take configured
table figure filter into account. As the table selector is now used
instead of the table figure to match the table element to its definition
(in most cases), we should make sure that the configured filter still
applies. Otherwise, in cases with multiple CALS table definitions, we
might pick the wrong one.